# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SEO Suite API
  version: "2.0.0"
  summary: SEO audit + indexing checker dashboard
  description: |
    REST + Server-Sent Events API for SEO Suite.

    **Versioning:** every path is reachable at both `/api/<path>` (current
    default) and `/api/v1/<path>` (versioned alias). Use the versioned form
    in new integrations so future breaking changes can ship as `/api/v2`
    without disrupting existing clients.

    **Auth:** session cookies issued by `/login`. All `/api/*` endpoints
    require `@login_required` unless noted otherwise. Some admin endpoints
    additionally require an admin flag.

    **Rate limiting:** per-IP, defaults to 240/minute. Stricter limits on
    `/login` (5/min), `/signup` (10/min), audit + indexing runs (10/hour).

    **SSRF:** every URL the server fetches goes through
    `validate_public_url()`, which rejects localhost / RFC1918 / link-local /
    metadata addresses and re-validates each redirect hop.

  contact:
    name: SEO Suite
    url: https://github.com/Surya8991/SEO-Suite
  license:
    name: MIT
    url: https://github.com/Surya8991/SEO-Suite/blob/main/LICENSE
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://seo-suite.example.com
    description: Production (replace with your domain)
...
paths:
...
  /login:
    get:
      tags: [Auth]
      summary: Login form (HTML)
      security: []
      responses:
        "200": {description: HTML login form}
    post:
      tags: [Auth]
      summary: Submit login credentials
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, password]
              properties:
                username: {type: string}
                password: {type: string, format: password}
                _csrf_token: {type: string}
                next: {type: string, description: "Path to redirect to after login"}
      responses:
        "302": {description: Login OK, redirected to /app or ?next path}
        "401": {description: Invalid credentials}
        "429": {description: Account temporarily locked or rate limit hit}
...
  /signup:
    post:
      tags: [Auth]
      summary: Self-register a new user
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, password, confirm]
              properties:
                username: {type: string, minLength: 3, maxLength: 64}
                password: {type: string, format: password, minLength: 12}
                confirm: {type: string, format: password}
                _csrf_token: {type: string}
      responses:
        "302": {description: Created + auto-logged-in, redirects to /app}
        "400": {description: Validation error}
        "403": {description: Signups disabled}
  /api/users:
    get:
      tags: [Auth]
      summary: List all users (admin only)
      responses:
        "200":
          description: User list
          content:
            application/json: {schema: {$ref: "#/components/schemas/UsersList"}}
        "401": {description: Not authenticated}
        "403": {description: Not admin}
    post:
      tags: [Auth]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: {type: string}
                password: {type: string, format: password}
                is_admin: {type: boolean, default: false}
      responses:
        "200": {description: Created}
        "400": {description: Validation error}
  /api/users/{username}:
    delete:
      tags: [Auth]
      summary: Delete a user (admin only)
      parameters:
        - in: path
          name: username
          required: true
          schema: {type: string}
      responses:
        "200": {description: Deleted}
        "400": {description: Cannot delete env admin or self}
  /api/auth/change_password:
    post:
      tags: [Auth]
      summary: Change your own password
      description: |
        Re-hashes the current user's password with argon2id after verifying
        the current password. The env admin (hash lives in
        `SEO_SUITE_PASSWORD_HASH`) is rejected — rotate that out-of-band.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: {type: string, format: password}
                new_password: {type: string, format: password, minLength: 12}
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: {type: boolean}
                  message: {type: string}
        "400": {description: "Policy violation (too short, same as current, etc.)"}
        "401": {description: Current password incorrect, or not logged in}
  /api/auth/login_history:
    get:
      tags: [Auth]
      summary: Recent login attempts (admin only)
      description: |
        Returns the most recent login attempts (successes and failures) from
        the SQLite `login_attempts` table. Records are retained for 90 days,
        then pruned at the next write.
      parameters:
        - in: query
          name: username
          schema: {type: string}
          description: Filter to one account
        - in: query
          name: failures_only
          schema: {type: string, enum: ["1", "true", "yes"]}
          description: Only failed attempts (for brute-force investigation)
        - in: query
          name: limit
          schema: {type: integer, minimum: 1, maximum: 500, default: 100}
      responses:
        "200":
          description: Login attempt rows, newest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: {type: boolean}
                  count: {type: integer}
                  rows: {type: array, items: {$ref: "#/components/schemas/LoginAttempt"}}
        "503": {description: SQLite backend disabled}
  /api/auth/my_logins:
    get:
      tags: [Auth]
      summary: Current user's own login history
      description: |
        Returns the last 50 login attempts for the authenticated user. Lets
        users spot unauthorized access without admin help.
      responses:
        "200":
          description: Login attempt rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: {type: boolean}
                  count: {type: integer}
                  rows: {type: array, items: {$ref: "#/components/schemas/LoginAttempt"}}
        "401": {description: Not logged in}
        "503": {description: SQLite backend disabled}

  # ── Audit ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    session_cookie:
      type: apiKey
      in: cookie
      name: session
      description: Flask session cookie set by /login.
  schemas:
...
    LoginAttempt:
      type: object
      properties:
        id: {type: integer}
        username: {type: string}
        ts: {type: string, format: date-time, description: "ISO-8601 UTC"}
        ip: {type: string, nullable: true}
        user_agent: {type: string, nullable: true, maxLength: 512}
        success: {type: boolean}
        reason:
          type: string
          nullable: true
          enum: [ok, env_admin, bad_password, unknown_user, locked_out]
          description: Machine-readable outcome code
    UsersList:
      type: object
      properties:
        ok: {type: boolean}
        users:
          type: array
          items:
            type: object
            properties:
              username: {type: string}
              is_admin: {type: boolean}
              created_at: {type: string, format: date-time}
        me: {type: string}
        signup_allowed: {type: boolean}
...
security:
  - session_cookie: []